+++ src/networkViz.ts
@@ -809,6 +809,11 @@ export default function networkVizJS(documentId: string, userLayoutOptions?: I.l
         layoutOptions.mouseOutNode = mouseOutCallback;
     }
 
+    // Function called when mousedown on node.
+    function setMouseDown(mouseDownCallback: any){
+        layoutOptions.mouseDownNode = mouseDownCallback;
+    }
+
     /**
      * Merges a node into another group.
      * If this node was in another group previously it removes it from the prior group.